services/kreditkarten/pyproject.toml
[project]
name = "kreditkarten-app"
version = "0.1.0"
description = "Web-Applikation zum Abgleich von Kreditkartenabrechnungen mit Belegen"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.2",
    "reportlab>=4.1.0",
    "cryptography>=42.0.0",
    "pytesseract>=0.3.10",
    "pillow>=10.2.0",
    "pdf2image>=1.17.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "markdown>=3.5.0",
    "flask-restx>=1.3.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]
